simplifiers: continued WIP
