+++ wowdump/simplifiers/__init__.py
@@ -8,9 +8,14 @@ import wowdump.simplifiers
 simplifiers = set()
 
 simplifier_list = frozenset([
-    "remove",
+    "bones",
     "color",
     "coordinates",
+    "enum",
+    "events",
+    "fileid",
+    "flags",
+    "remove",
     "shaderid_m2",
     "shaderid_wmo",
 ])
@@ -21,4 +26,24 @@ for s in simplifier_list:
         compiled_re = re.compile(sss[0], re.VERBOSE)
         simplifiers.add((compiled_re, sss[1]))
 
-print(simplifiers)
+
+# Check to see if a given path has a simplifier, return the appropriate
+# simplifier function if there's a match.
+def check_simplify(path: str):
+    for r in simplifiers:
+        if r[0].search(path):
+            return r[1]
+
+    return None
+
+
+# Notes from the original implementation:
+#
+# These next bits are an attempt to simplify certain structures into
+# a more human-readable form. This is ugly and stupid and needs to
+# be hand-maintained. We might be able to do better with some additional
+# metadata out of kaitai or such, but for now, this is a quick and dirty
+# "but I want to be able to read this" solution. There's probably better
+# bad ways to do it, and definitely better not-bad ways to do it, but
+# this works. For now. Until it offends my sight just a little too much
+# and gets torn out by the roots (just like my hair)   --A
