+++ wowdump/simplifiers/fileid.py
@@ -1,0 +1,75 @@
+import sys
+from typing import Any, Optional, Dict, List, Callable
+
+# We can run without, we'll just be slow
+try:
+    import sqlite3
+except ImportError:
+    pass
+
+
+# simplify file ids of various types
+def simplify_fileid(id: id, _parent, cachecon) -> str:
+    # FIXME: we probably need to pass args to simplifiers
+    # if not args.resolve or id <= 0:
+    #     return f"{id}"
+    if id <= 0:
+        return f"{id}"
+
+    c = cache_getfileid(id, cachecon)
+
+    # if None, cache is not usable, so try the old fashioned way
+    if c is None:
+        c = csv_getfileid(id)
+
+    if c is False:
+        return f"{id}  # unresolved"
+    else:
+        return f"{id}  # {c}"
+
+
+# FIXME: Should this be somewhere else?
+def cache_getfileid(id: int, cachecon) -> Optional[str]:
+    if not cachecon:
+        return None
+
+    cur = cachecon.cursor()
+    q = "SELECT name FROM file_ids WHERE id=?"
+
+    try:
+        res = cur.execute(q, [id]).fetchone()
+    except sqlite3.Error as e:
+        print(f"ERROR: Unexpected sqlite error: {e}", file=sys.stderr)
+        return False
+
+    # not in cache? Return false so that we know not to try to fall back to
+    # the listfile, since the cache should be authoritative
+    if res is None:
+        return False
+
+    return res[0]
+
+
+# FIXME: Where should we -actually- look for this file?
+def csv_getfileid(id: int):
+    try:
+        with open("listfile.csv", 'r', newline="") as csvfile:
+            reader = csv.reader(csvfile, delimiter=";")
+            for row in reader:
+                if int(row[0]) == id:
+                    return f"{id}  # {row[1]}"
+    except:
+        pass
+
+    return False
+
+
+fileids_re = r"^/chunks/\d+/chunk_data/([^/]+_)?file_data_ids/\d+$"
+fileid_re = r"/([^/]+_)?file_data_id$"
+mapfileid_re = r"^/chunks/\d+/chunk_data/map_fileids/\d+/[^/]+_file_data_id$"
+
+simplifiers = [
+    (fileid_re, simplify_fileid),
+    (fileids_re, simplify_fileid),
+    (mapfileid_re, simplify_fileid),
+]
