+++ wowdump/wowdump.py
@@ -12,7 +12,7 @@ from ppretty import ppretty
 import inspect
 
 from kaitaistruct import KaitaiStruct, KaitaiStream, BytesIO
-from . import simplifiers
+from .simplifiers import check_simplify
 
 # We can run without, we'll just be slow
 try:
@@ -231,6 +231,7 @@ def to_tree(obj, path: str = ""):
 
 # Caching bits (yeah, they're ugly)
 # or maybe type of os.PathLike for cache_open
+# FIXME: move caching bits to separate module
 def cache_open(dbfile: str):
     if "sqlite3" not in sys.modules:
         print("WARNING: sqlite not available, caching disabled", file=sys.stderr)
@@ -274,38 +275,10 @@ def cache_fileids(listfile: str, cachecon) -> None:
         f"INFO: fileid cache successfully rebuilt in {runtime:.2f}s", file=sys.stderr)
 
 
-def cache_getfileid(id: int, cachecon) -> Optional[str]:
-    if not cachecon:
-        return None
-
-    cur = cachecon.cursor()
-    q = "SELECT name FROM file_ids WHERE id=?"
-
-    try:
-        res = cur.execute(q, [id]).fetchone()
-    except sqlite3.Error as e:
-        print(f"ERROR: Unexpected sqlite error: {e}", file=sys.stderr)
-        return False
-
-    # not in cache? Return false so that we know not to try to fall back to
-    # the listfile, since the cache should be authoritative
-    if res is None:
-        return False
-
-    return res[0]
 
 
-def csv_getfileid(id: int):
-    try:
-        with open("listfile.csv", 'r', newline="") as csvfile:
-            reader = csv.reader(csvfile, delimiter=";")
-            for row in reader:
-                if int(row[0]) == id:
-                    return f"{id}  # {row[1]}"
-    except:
-        pass
 
-    return False
+
 
 
 # There's probably a way better way to do this.
@@ -369,79 +342,12 @@ def check_filtered(path: str) -> bool:
     return False
 
 
-# These next bits are an attempt to simplify certain structures into
-# a more human-readable form. This is ugly and stupid and needs to
-# be hand-maintained. We might be able to do better with some additional
-# metadata out of kaitai or such, but for now, this is a quick and dirty
-# "but I want to be able to read this" solution. There's probably better
-# bad ways to do it, and definitely better not-bad ways to do it, but
-# this works. For now. Until it offends my sight just a little too much
-# and gets torn out by the roots (just like my hair)   --A
-
-flags_re = re.compile(r'''
-    /(global_)?flags$
-''', re.VERBOSE)
-
-
-
-fileids_re = re.compile(r'''
-    ^/chunks/\d+/chunk_data/([^/]+_)?file_data_ids/\d+$
-''', re.VERBOSE)
-
-fileid_re = re.compile(r'''
-    /([^/]+_)?file_data_id$
-''', re.VERBOSE)
-
-# FIXME: try to combine with the above
-mapfileid_re = re.compile(r'''
-    ^/chunks/\d+/chunk_data/map_fileids/\d+/[^/]+_file_data_id$
-''', re.VERBOSE)
-
-version_re = re.compile(r'''
-    ^/model/version$
-''', re.VERBOSE)
-
-interpolation_type_re = re.compile(r'''
-    interpolation_type$
-''', re.VERBOSE)
-
-
-
-# simplify_4bone
-fourbone_re = re.compile(r'''
-    ^/model/vertices/\d+/(bone_indices|bone_weights)$
-''', re.VERBOSE)
-
-
-eventid_re = re.compile(r'''
-    ^/model/events/\d+/eventid$
-''', re.VERBOSE)
 
 
 
 
 
 
-
-
-def simplify_fourbone(d, _parent, _cachecon) -> str:
-    return f"[ {d[0]}, {d[1]}, {d[2]}, {d[3]} ]"
-
-
-# FIXME: Should the output be inside { } or something?
-def simplify_flags(d, _parent, _cachecon):
-    if not isinstance(d, dict):
-        return d
-
-    flags = []
-    for k, v in d.items():
-        if v:
-            flags.append(k)
-
-    if len(flags) == 0:
-        return "none"
-
-    return ", ".join(flags)
 
 
 interpolation_types = {
@@ -451,8 +357,6 @@ interpolation_types = {
     3: "interpolate_cubic_hermite_spline",
 }
 
-def simplify_enum(d, _parent, _cachecon):
-    return f"{d['value']}  # {d['name']}"
 
 
 
@@ -463,190 +367,11 @@ def simplify_enum(d, _parent, _cachecon):
 
 
 
-m2_events = {
-    "$AH0": "PlaySoundKit (customAttack[0])",
-    "$AH1": "PlaySoundKit (customAttack[1])",
-    "$AH2": "PlaySoundKit (customAttack[2])",
-    "$AH3": "PlaySoundKit (customAttack[3])",
-    "$BMD": "BowMissleDestination",
-    "$ALT": "Anim Swap Event",
-    "$BL0": "Backwards Footstep Anim Event Hit L",
-    "$BL1": "Backwards Footstep Anim Event Hit L",
-    "$BL2": "Backwards Footstep Anim Event Hit L",
-    "$BL3": "Backwards Footstep Anim Event Hit L",
-    "$BR0": "Backwards Footstep Anim Event Hit R",
-    "$BR1": "Backwards Footstep Anim Event Hit R",
-    "$BR2": "Backwards Footstep Anim Event Hit R",
-    "$BR3": "Backwards Footstep Anim Event Hit R",
-    # soundEffect ID is defined by CreatureSoundDatarec::m_birthSoundID
-    "$BRT": "PlaySoundKit (birth)",
-    "$BTH": "Breath",
-    "$BWP": "PlayRangedItemPull (Bow Pull)",
-    "$BWR": "BowRelease",
-    "$CAH": "Attack Hold",
-    # Only z is used. Non-animated. Not used if $CMA
-    "$CFM": "CGCamera::UpdateMountHeightOrOffset",
-    "$CMA": "CGCamera::UpdateMountHeightOrOffset",  # Position for camera
-    # parry, anims, depending on some state, also some callback which might do more
-    "$CPP": "PlayCombatActionAnimKit",
-    "$CSD": "PlayEmoteSound",  # data = soundEntryId
-    "$CSL": "release missles if pending (L)",
-    "$CSR": "release missles if pending (R)",
-    # sound played depends on CGUnit_C::GetWeaponSwingType
-    "$CSS": "PlayWeaponSwooshSound",
-    "$CST": "release missles if pending",
-    "$DSE": "DestroyEmitter",
-    "$DSO": "DoodadSoundOneShot",
-    "$DTH": "DeathThud + LootEffect",
-    "$EAC": "object package state enter 3, exit 2, 4, 5",
-    "$EDC": "object package state enter 5, exit 3, 4, 2",
-    "$EMV": "object package state enter 4, exit 3, 2, 5",
-    # soundEffect ID is implicit by currently played emote
-    "$ESD": "PlayEmoteStateSound",
-    "$EWT": "object package state enter 2, exit 3, 4, 5",
-    "$FD0": "PlayFidgetSound[0]",  # from CreatureSoundDataRec::m_soundFidget
-    "$FD1": "PlayFidgetSound[1]",
-    "$FD2": "PlayFidgetSound[2]",
-    "$FD3": "PlayFidgetSound[3]",
-    "$FD4": "PlayFidgetSound[4]",
-    "$FD5": "PlayFidgetSound[5]",
-    "$FD6": "PlayFidgetSound[6]",
-    "$FD7": "PlayFidgetSound[7]",
-    "$FD8": "PlayFidgetSound[8]",
-    "$FD9": "PlayFidgetSound[9]",
-    # soundEffect ID is defined by CreatureSoundDataRec::m_soundStandID
-    "$FDX": "PlayUnitSound(stand)",
-    "$FL0": "FootstepAnimEventHitL[0]",
-    "$FL1": "FootstepAnimEventHitL[1]",
-    "$FL2": "FootstepAnimEventHitL[2]",
-    "$FL3": "FootstepAnimEventHitL[3]",
-    "$FR0": "FootstepAnimEventHitR[0]",
-    "$FR1": "FootstepAnimEventHitR[1]",
-    "$FR2": "FootstepAnimEventHitR[2]",
-    "$FR3": "FootstepAnimEventHitR[3]",
-    "$FSD": "HandleFootfallAnimEvent",
-    # soundEffect ID is defined by GameObjectDisplayInfoRec::m_Sound[x + 6] ({Custom0, Custom1, Custom2, Custom3})
-    "$GC0": "GameObject_C_PlayAnimatedSound[0]",
-    "$GC1": "GameObject_C_PlayAnimatedSound[1]",
-    "$GC2": "GameObject_C_PlayAnimatedSound[2]",
-    "$GC3": "GameObject_C_PlayAnimatedSound[3]",
-    # soundEffect ID is defined by GameObjectDisplayInfoRec::m_Sound[x] ({Stand, Open, Loop, Close, Destroy, Opened})
-    "$GO0": "GameObject_C_PlayAnimatedSound2[0]",
-    "$GO1": "GameObject_C_PlayAnimatedSound2[1]",
-    "$GO2": "GameObject_C_PlayAnimatedSound2[2]",
-    "$GO3": "GameObject_C_PlayAnimatedSound2[3]",
-    "$GO4": "GameObject_C_PlayAnimatedSound2[4]",
-    "$GO5": "GameObject_C_PlayAnimatedSound2[5]",
-    "$HIT": "PlayWoundAnimKit",  # soundEntryId depends on SpellVisualKit
-    "$RL0": "FootstepAnimEventHitL[0] (running)",
-    "$RL1": "FootstepAnimEventHitL[1] (running)",
-    "$RL2": "FootstepAnimEventHitL[2] (running)",
-    "$RL3": "FootstepAnimEventHitL[3] (running)",
-    "$RR0": "FootstepAnimEventHitR[0] (running)",
-    "$RR1": "FootstepAnimEventHitR[1] (running)",
-    "$RR2": "FootstepAnimEventHitR[2] (running)",
-    "$RR3": "FootstepAnimEventHitR[3] (running)",
-    # soundEffect ID is defined by CreatureSoundDataRec::m_spellCastDirectedSoundID
-    "$SCD": "PlaySoundKit (spellCastDirectedSound)",
-    "$SHK": "AddShake",  # arg is spellEffectCameraShakesID
-    "$SHL": "ExchangeSheathedWeaponL",
-    "$SHR": "ExchangeSheathedWeaponR",
-    "$SL0": "FootstepAnimEventHitL[0]",
-    "$SL1": "FootstepAnimEventHitL[1]",
-    "$SL2": "FootstepAnimEventHitL[2]",
-    "$SL3": "FootstepAnimEventHitL[3]",
-    # soundEffect ID is defined by CreatureSoundDatarec::m_submergedSoundID
-    "$SMD": "PlaySoundKit (submerged)",
-    # soundEffect ID is defined by CreatureSoundDatarec::m_submergeSoundID
-    "$SMG": "PlaySoundKit (submerge)",
-    "$SND": "PlaySoundKit (custom)",  # arg is soundEntryId
-    "$SR0": "FootstepAnimEventHitR[0]",
-    "$SR1": "FootstepAnimEventHitR[1]",
-    "$SR2": "FootstepAnimEventHitR[2]",
-    "$SR3": "FootstepAnimEventHitR[3]",
-    # Not seen in 6.0.1.18179 -- x is {E and B} , sequence time is taken of both, pivot of $STB. (Also, attachment info for attachment 0)
-    "$STx": "Mount(?)",
-    "$TRD": "HandleSpellEventSound",  # soundEffect ID is implicit by SpellRec
-    "$VG0": "HandleBoneAnimGrabEvent[0]",
-    "$VG1": "HandleBoneAnimGrabEvent[1]",
-    "$VG2": "HandleBoneAnimGrabEvent[2]",
-    "$VG3": "HandleBoneAnimGrabEvent[3]",
-    "$VG4": "HandleBoneAnimGrabEvent[4]",
-    "$VG5": "HandleBoneAnimGrabEvent[5]",
-    "$VG6": "HandleBoneAnimGrabEvent[6]",
-    "$VG7": "HandleBoneAnimGrabEvent[7]",
-    "$VG8": "HandleBoneAnimGrabEvent[8]",
-    "$VT0": "HandleBoneAnimThrowEvent[0]",
-    "$VT1": "HandleBoneAnimThrowEvent[1]",
-    "$VT2": "HandleBoneAnimThrowEvent[2]",
-    "$VT3": "HandleBoneAnimThrowEvent[3]",
-    "$VT4": "HandleBoneAnimThrowEvent[4]",
-    "$VT5": "HandleBoneAnimThrowEvent[5]",
-    "$VT6": "HandleBoneAnimThrowEvent[6]",
-    "$VT7": "HandleBoneAnimThrowEvent[7]",
-    "$VT8": "HandleBoneAnimThrowEvent[8]",
-    # soundEffect ID is defined by CreatureSoundDataRec::m_soundWingGlideID
-    "$WGG": "PlayUnitSound (wingGlide)",
-    # soundEffect ID is defined by CreatureSoundDataRec::m_soundWingFlapID
-    "$WNG": "PlayUnitSound (wingFlap)",
-    "$WR0": "FootstepAnimEventHitL[0]",
-    "$WR1": "FootstepAnimEventHitL[1]",
-    "$WR2": "FootstepAnimEventHitL[2]",
-    "$WR3": "FootstepAnimEventHitL[3]",
-    "$WRB": "Bow Something",
-    "$WRT": "Bow Something",
-}
-
-def simplify_events(d, _parent, _cachecon):
-    if d in m2_events:
-        return f"{d}  # {m2_events[d]}"
-    else:
-        return d
-
-
-def simplify_fileid(id: id, _parent, cachecon) -> str:
-    if not args.resolve or id <= 0:
-        return f"{id}"
-
-    c = cache_getfileid(id, cachecon)
-
-    # if None, cache is not usable, so try the old fashioned way
-    if c is None:
-        c = csv_getfileid(id)
-
-    if c is False:
-        return f"{id}  # unresolved"
-    else:
-        return f"{id}  # {c}"
-
-
-simplifications = [
-
-    (flags_re, simplify_flags),
 
 
 
-    (fileid_re, simplify_fileid),
-    (fileids_re, simplify_fileid),
-    (mapfileid_re, simplify_fileid),
-    (interpolation_type_re, simplify_enum),
-    (fourbone_re, simplify_fourbone),
-    (version_re, simplify_enum),
 
 
-
-    (eventid_re, simplify_events),
-]
-
-def check_simplify(path: str):
-    if not args.simplify:
-        return None
-
-    for r in simplifications:
-        if r[0].search(path):
-            return r[1]
-
-    return None
 
 
 # Given a path, check to see if it's one we want to normally elide (because
@@ -695,7 +420,9 @@ def pathdump(d, path: str, cachecon) -> None:
         # if check_filtered(workpath):
         #     continue
 
-        s = check_simplify(workpath)
+        # doublecheck to make sure check_simply doesn't get called if disabled
+        s = check_simplify(workpath) if args.simplify else None
+
         if s:
             # if it's simplified, we're at a 'final' path, so check filtering
             if check_filtered(workpath):
